refactor: move Traits/UserProvider into UserModel
